Cucumber tests with 91% coverage.
