--- features/artists_ui.feature
@@ -1,8 +1,0 @@
-Feature: Empty Artists UI
-
-  Background:
-    Given the database is clean
-
-  Scenario: Artists index shows empty state
-    When I go to the artists page
-    Then I should see "No artists found yet"
