+++ features/home_controller.feature
@@ -1,4 +1,5 @@
-Feature: HomeController basic response
+Feature: HomeController works
+
   Scenario: Home index renders
     When I visit the home controller page
     Then I should see "VinylVerse"
