--- features/step_definitions/coverage.rb
@@ -1,150 +1,0 @@
-def swallow
-  yield
-rescue StandardError
-end
-
-module CukesFakeDiscogs
-  class Wrapper
-    def initialize(*); end
-
-    def search(query, type: :artist)
-      OpenStruct.new(
-        results: [
-          OpenStruct.new(title: "#{query} — Best Of", id: 101, type: type, country: 'UK', style: ['Rock']),
-          OpenStruct.new(title: "#{query} — Live",    id: 102, type: type, country: 'US', style: ['Pop'])
-        ]
-      )
-    end
-
-    def get_release(id)
-      case id.to_i
-      when 1, 101
-        { 'genres' => ['Rock'], 'country' => 'UK' }
-      when 2, 102
-        { 'genres' => ['Pop'],  'country' => 'US' }
-      else
-        {} 
-      end
-    end
-
-    def artist_releases(_artist_id)
-      {
-        'releases' => [
-          { 'id' => 1, 'title' => 'Alpha' },
-          { 'id' => 2, 'title' => 'Beta' }
-        ]
-      }
-    end
-  end
-end
-
-Given('I exercise backend code paths for coverage') do
-  rh = Rails.application.routes.url_helpers
-  swallow { rh.root_path }
-  swallow { rh.artists_path }
-  swallow { rh.new_artist_path }
-  swallow { rh.search_path } if rh.respond_to?(:search_path)
-
-  a_invalid = Artist.new(name: nil)
-  a_invalid.valid?
-  a_invalid.errors.full_messages.join(', ')
-
-  a = Artist.create!(name: 'Coverage Artist')
-  a.update(name: 'Coverage Artist Renamed')
-  a.name = ''
-  a.valid?
-  a.errors.full_messages.join(', ')
-  a.update(name: 'Final Artist Name')
-
-  if defined?(Album)
-    alb = (Album.where(title: 'Cuke Album').first ||
-           swallow { Album.create!(title: 'Cuke Album') } ||
-           Album.new(title: 'Cuke Album'))
-
-    if alb.respond_to?(:valid?)
-      alb.valid?
-      swallow { alb.errors.full_messages.join(', ') }
-    end
-  end
-
-  if defined?(CollectionItem)
-    ci = CollectionItem.new
-    ci.valid?
-    ci.errors.full_messages.join(', ')
-  end
-
-  if defined?(WishlistItem)
-    wi = WishlistItem.new
-    wi.valid?
-    wi.errors.full_messages.join(', ')
-  end
-
-  album = nil
-  swallow do
-    base_artist = Artist.where(name: 'Items Artist').first || Artist.create!(name: 'Items Artist')
-    album = if defined?(Album)
-      Album.where(title: 'Items Album').first || Album.create!(title: 'Items Album', artist: base_artist) rescue nil
-    end
-  end
-
-  swallow do
-    CollectionItem.create!(album: album) if defined?(CollectionItem) && album
-    WishlistItem.create!(album: album)   if defined?(WishlistItem)   && album
-  end
-
-  swallow { ApplicationController.helpers.inspect }
-
-  original_discogs = Object.const_get(:Discogs) if Object.const_defined?(:Discogs)
-  begin
-    Object.send(:remove_const, :Discogs) if Object.const_defined?(:Discogs)
-    Object.const_set(:Discogs, Module.new)
-    Discogs.const_set(:Wrapper, CukesFakeDiscogs::Wrapper)
-
-    service_path = Rails.root.join('app', 'services', 'discogs_service.rb')
-    swallow { load service_path if File.file?(service_path) }
-
-    if defined?(DiscogsService)
-      svc = DiscogsService.new
-
-      swallow { svc.search_artist('The Beatles') }
-
-      swallow do
-        if svc.respond_to?(:count_releases_for_artist_id)
-          svc.count_releases_for_artist_id(82730)
-        end
-      end
-
-      swallow do
-        if svc.respond_to?(:genre_for_artist)
-          svc.genre_for_artist(82730)
-        end
-      end
-
-      swallow do
-        if svc.respond_to?(:country_for_artist)
-          svc.country_for_artist(82730)
-        end
-      end
-
-      swallow { svc.send(:get_release, 999_999) if svc.respond_to?(:get_release, true) }
-    end
-  ensure
-    Object.send(:remove_const, :Discogs) if Object.const_defined?(:Discogs)
-    Object.const_set(:Discogs, original_discogs) if defined?(original_discogs) && original_discogs
-  end
-
-  rh = Rails.application.routes.url_helpers
-  if Capybara.current_session.driver.respond_to?(:submit)
-    swallow { page.driver.submit :post, rh.artists_path, { artist: { name: '' } } }
-
-    swallow do
-      page.driver.submit :post, rh.artists_path, { artist: { name: 'Temp Cuke Artist' } }
-      created = Artist.where(name: 'Temp Cuke Artist').order('id DESC').first
-      if created
-        page.driver.submit :delete, rh.artist_path(created), {}
-      end
-    end
-  end
-
-  expect(true).to be true
-end
