--- features/step_definitions/signup_steps.rb
@@ -1,4 +1,0 @@
-Given('I am on the signup page') do
-  visit '/signup'
-end
-
